@@ -8,21 +8,24 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@inkline/inkline": "^3.0.8",
-    "@mdi/font": "^6.7.96",
+    "@mdi/font": "^7.0.96",
+    "@vicons/ionicons5": "^0.12.0",
+    "@vueuse/core": "^9.1.0",
     "axios": "^0.27.2",
-    "eslint": "^8.16.0",
+    "eslint": "^8.21.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-vue": "^9.0.1",
-    "pinia": "^2.0.14",
-    "vue": "^3.2.25",
-    "vue-router": "^4.0.15"
+    "eslint-plugin-vue": "^9.3.0",
+    "mitt": "^3.0.0",
+    "pinia": "^2.0.22",
+    "vue": "^3.2.39",
+    "vue-router": "^4.1.5"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.3",
-    "sass": "^1.52.1",
-    "vite": "^2.9.9",
-    "vite-plugin-fonts": "^0.4.0"
+    "@vitejs/plugin-vue": "^3.0.1",
+    "naive-ui": "^2.33.3",
+    "sass": "^1.54.3",
+    "vite": "^3.1.3",
+    "vite-plugin-fonts": "^0.6.0"
   }
 }
