@@ -1,19 +1,17 @@
 import { createApp } from 'vue';
-import { Inkline, components } from '@inkline/inkline';
 import { createPinia } from 'pinia';
+import naive from 'naive-ui';
+import emitter from './emitter';
 import App from './App.vue';
 import router from './router';
 
-import '@inkline/inkline/inkline.scss';
-import './main.scss';
+import './variables.scss';
 // eslint-disable-next-line import/no-unresolved
 import 'virtual:fonts.css';
 
 createApp(App)
-  .use(Inkline, {
-    components,
-    colorMode: 'light',
-  })
+  .use(naive)
   .use(createPinia())
   .use(router)
+  .use(emitter)
   .mount('#app');
